move code to a capability class
